refactor(api): update campaign brief validation and refactor variant generation by moving logic to a dedicated service to call for both cli and fastapi
- Change validation to require at least 1 product in CampaignBrief
- Update API endpoints for better readability and maintainability
- Adjust example JSON and documentation for clarity
