+++ src/cli.py
@@ -10,11 +10,12 @@ import json
 import sys
 from pathlib import Path
 
-from src.models.brief import ASPECT_RATIOS, AspectRatio, CampaignBrief, Product
+from src.models.brief import ASPECT_RATIOS, CampaignBrief
 from src.services.genai import GenAIOrchestrator
 from src.services.openai_image_client import OpenAIImageClient
 from src.services.processor import ImageProcessor
 from src.services.storage import StorageManager
+from src.services.variant_generation import generate_variant
 from src.utils.config import settings
 from src.utils.logger import get_logger
 
@@ -134,144 +135,6 @@ async def process_campaign(brief_path: str) -> None:
     logger.info("=" * 80)
 
 
-async def generate_variant(
-    orchestrator: GenAIOrchestrator,
-    processor: ImageProcessor,
-    storage: StorageManager,
-    product: Product,
-    brief: CampaignBrief,
-    ratio: AspectRatio,
-) -> dict:
-    """Generate single variant for a product at specific aspect ratio
-
-    Workflow:
-    1. Check for existing asset (reuse if available)
-    2. Generate new image via GenAI if needed
-    3. Resize to target aspect ratio
-    4. Add campaign message text overlay
-    5. Save to organized folder structure
-
-    Args:
-        orchestrator: GenAI orchestration service
-        processor: Image processing service
-        storage: Storage management service
-        product: Product information
-        brief: Campaign brief
-        ratio: Target aspect ratio
-
-    Returns:
-        Dictionary with generation result metadata
-    """
-    reused = False
-
-    try:
-        # Step 1: Check if asset exists
-        existing_asset = storage.get_asset(
-            product_id=product.id, ratio_name=ratio.name, campaign_id=brief.campaign_id
-        )
-
-        if existing_asset:
-            logger.debug(f"Reusing existing asset: {product.id}/{ratio.name}")
-            image_data = existing_asset
-            reused = True
-        else:
-            # Step 2: Generate new image
-            prompt = build_generation_prompt(product, brief)
-            logger.debug(f"Prompt for {product.id}: {prompt[:100]}...")
-
-            image_data = await orchestrator.generate_image(
-                prompt=prompt,
-                width=ratio.width,
-                height=ratio.height,
-                product_id=product.id,
-            )
-
-            # Step 3: Resize to exact target dimensions
-            image_data = processor.resize(image_data, ratio.width, ratio.height)
-
-        # Step 4: Add text overlay (always applied, even for reused assets)
-        image_data = processor.add_text_overlay(
-            image_data=image_data, text=brief.campaign_message, position="bottom"
-        )
-
-        # Step 5: Save output with metadata
-        output_path = storage.save_output(
-            product_id=product.id,
-            ratio_name=ratio.name,
-            image_data=image_data,
-            campaign_id=brief.campaign_id,
-            metadata={
-                "campaign_id": brief.campaign_id,
-                "product_id": product.id,
-                "product_name": product.name,
-                "aspect_ratio": ratio.name,
-                "dimensions": f"{ratio.width}x{ratio.height}",
-                "platform": ratio.platform,
-                "target_market": brief.target_market,
-                "target_audience": brief.target_audience,
-                "campaign_message": brief.campaign_message,
-                "reused": reused,
-            },
-        )
-
-        return {
-            "success": True,
-            "ratio": ratio.name,
-            "path": str(output_path),
-            "reused": reused,
-        }
-
-    except Exception as e:
-        logger.error(f"Failed to generate {ratio.name} for {product.id}: {e}")
-        raise
-
-
-def build_generation_prompt(product: Product, brief: CampaignBrief) -> str:
-    """Build optimised prompt for GenAI image generation
-
-    Prompt engineering strategy:
-    - Product name and description
-    - Target audience context
-    - Market/locale hints
-    - Style and quality keywords
-
-    Args:
-        product: Product information
-        brief: Campaign brief
-
-    Returns:
-        Optimised generation prompt
-    """
-    parts = [product.name,]
-
-    if product.description:
-        parts.append(product.description)
-
-    # Add audience and market context
-    parts.extend(
-        [
-            f"for {brief.target_audience}",
-            f"in {brief.target_market} market",
-        ]
-    )
-
-    # Add style keywords
-    parts.extend(
-        [
-            "professional product photography",
-            "high quality",
-            "marketing campaign",
-            "clean background",
-        ]
-    )
-
-    prompt = ", ".join(parts)
-
-    # Truncate if too long (OpenAI image generation has ~4000 characters limit)
-    if len(prompt) > 1000:
-        prompt = prompt[:1000]
-
-    return prompt
 
 
 def main():
